+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add information about superuser privileges when creating API tokens
 * Migrate documentation from mkdocs to vitepress
 * Support multi select and bulk actions in report sidebar
+* Support bulk create findings from templates
 * Highlight inactive users in project member selection
 * Improve handling of short search terms in project search